@@ -1,6 +1,7 @@
 package fsm
 
 import (
+	"github.com/vitaliy-ukiru/fsm-telebot/internal"
 	tele "gopkg.in/telebot.v3"
 )
 
@@ -16,8 +17,9 @@ type Manager struct {
 	bot          *tele.Bot
 	group        *tele.Group // handlers will add to group
 	store        Storage
-	handlers     handlerStorage
+	handlers     handlerMapping
 	contextMaker ContextMakerFunc
+	g            []tele.MiddlewareFunc
 }
 
 // NewManager returns new Manger.
@@ -38,7 +40,7 @@ func NewManager(
 		group:        group,
 		store:        storage,
 		contextMaker: ctxMaker,
-		handlers:     make(handlerStorage),
+		handlers:     make(handlerMapping),
 	}
 }
 
@@ -57,6 +59,7 @@ func (m *Manager) With(g *tele.Group) *Manager {
 		store:        m.store,
 		handlers:     m.handlers,
 		contextMaker: m.contextMaker,
+		g:            m.g,
 	}
 }
 
@@ -65,22 +68,25 @@ func (m *Manager) SetContextMaker(contextMaker ContextMakerFunc) {
 	m.contextMaker = contextMaker
 }
 
-// NewGroup returns Manager copy with new tele.Group.
-//
-// Deprecated: Incorrect behavior with separated groups.
+// NewGroup returns manager child with copy
+// of middleware group. Adding middlewares in
+// new group doesn't affect the parent.
 func (m *Manager) NewGroup() *Manager {
+	g := make([]tele.MiddlewareFunc, len(m.g))
+	copy(g, m.g)
 	return &Manager{
 		bot:          m.bot,
-		group:        m.bot.Group(),
+		group:        m.group,
 		store:        m.store,
 		handlers:     m.handlers,
 		contextMaker: m.contextMaker,
+		g:            g,
 	}
 }
 
 // Use add middlewares to group.
 func (m *Manager) Use(middlewares ...tele.MiddlewareFunc) {
-	m.group.Use(middlewares...)
+	m.g = append(m.g, middlewares...)
 }
 
 // Bind adds handler (with FSM context argument) with filter on state.
@@ -88,8 +94,8 @@ func (m *Manager) Use(middlewares ...tele.MiddlewareFunc) {
 // Difference between Bind and Handle methods what Handle require Filter objects.
 // And this method can work with only one state.
 // If you bind some states see docs to Handle.
-func (m *Manager) Bind(end interface{}, state State, h Handler, middlewares ...tele.MiddlewareFunc) {
-	m.Handle(F(end, state), h, middlewares...)
+func (m *Manager) Bind(end any, state State, h Handler, middlewares ...tele.MiddlewareFunc) {
+	m.handle(end, []State{state}, h, middlewares)
 }
 
 // Handle adds handler to group chain with filter on states.
@@ -98,7 +104,7 @@ func (m *Manager) Bind(end interface{}, state State, h Handler, middlewares ...t
 // Binding some states to one handler.
 //
 //	var ( // types of variables
-//		endpoint interface{} // string | tele.CallbackEndpoint
+//		endpoint any // string | tele.CallbackEndpoint
 //		states []State
 //		handlerFunc fsm.Handler
 //	)
@@ -106,22 +112,66 @@ func (m *Manager) Bind(end interface{}, state State, h Handler, middlewares ...t
 //	// or
 //	manager.Handle(fsm.Filter{endpoint, states}, handlerFunc)
 func (m *Manager) Handle(f Filter, h Handler, middlewares ...tele.MiddlewareFunc) {
-	endpoint := f.CallbackUnique()
 	if len(f.States) == 0 {
 		f.States = []State{DefaultState}
 	}
 
-	m.handlers.add(endpoint, h, f.States)
+	m.handle(f.Endpoint, f.States, h, middlewares)
+}
+
+func (m *Manager) handle(
+	end any,
+	states []State,
+	h Handler,
+	ms []tele.MiddlewareFunc,
+) {
+	endpoint := getEndpoint(end)
+
+	// we handles multi handlers in telebot,
+	// so need to use middleware here
+	wrappedHandler := m.withMiddleware(m.adapter(h), ms)
+	m.handlers.add(endpoint, wrappedHandler, states)
+
 	m.group.Handle(
 		endpoint,
-		m.HandlerAdapter(m.handlers.forEndpoint(endpoint)),
-		middlewares...,
+		m.forEndpoint(endpoint),
 	)
 }
 
-// HandlerAdapter create telebot.HandlerFunc object for Handler with FSM context.
+// withMiddleware join handler middlewares with group middlewares.
+func (m *Manager) withMiddleware(h tele.HandlerFunc, ms []tele.MiddlewareFunc) tele.HandlerFunc {
+	ms = append(m.g, ms...)
+
+	// I didn’t understand why ApplyMiddleware is called
+	// inside the handler, just copied from telebot code.
+	return func(c tele.Context) error {
+		return internal.ApplyMiddleware(h, ms)(c)
+	}
+}
+
+// HandlerAdapter create telebot.HandlerFunc object
+// for Handler with FSM context.
+//
+// Used for external purposes only outside handlers chain.
+// Example: access to context without manager handlers.
+// Use only as directed and if you know what you are doing.
 func (m *Manager) HandlerAdapter(handler Handler) tele.HandlerFunc {
 	return func(c tele.Context) error {
+		return handler(c, m.contextMaker(c, m.store))
+	}
+}
+
+// adapter wraps internal Handler to telebot.
+// difference between HandlerAdapter in support
+// wrap context.
+func (m *Manager) adapter(handler Handler) tele.HandlerFunc {
+	return func(c tele.Context) error {
+		fsmCtx, ok := tryUnwrapContext(c)
+		if ok {
+			return handler(c, fsmCtx)
+		}
+
+		// bad case, creating new context
 		return handler(c, m.contextMaker(c, m.store))
 	}
 }
