@@ -1,15 +1,14 @@
 package fsm
 
 import (
-	"container/list"
 	"fmt"
 
 	"github.com/vitaliy-ukiru/fsm-telebot/internal"
 	tele "gopkg.in/telebot.v3"
 )
 
-// handlerStorage contains handlers group separated by endpoint.
-type handlerStorage map[string]*list.List
+// handlerMapping contains handlers group separated by endpoint.
+type handlerMapping map[string]*internal.List[handlerEntry]
 
 // handlerEntry representation handler with states, needed for add endpoints correct
 // Because telebot uses rule: 1 endpoint = 1 handler.
@@ -17,43 +16,57 @@ type handlerStorage map[string]*list.List
 //
 // We can use switch-case in handler for check states, but I think not best practice.
 type handlerEntry struct {
-	states  statesHashset
-	handler Handler
+	states  internal.HashSet[State]
+	handler tele.HandlerFunc
 }
 
 // add handler to storage, just shortcut.
-func (m handlerStorage) add(endpoint string, h Handler, states []State) {
-	statesSet := newHashsetFromSlice(states)
-	m.insert(endpoint, handlerEntry{states: statesSet, handler: h})
+func (hm handlerMapping) add(endpoint string, h tele.HandlerFunc, states []State) {
+	statesSet := internal.HashSetFromSlice(states)
+	hm.insert(endpoint, handlerEntry{states: statesSet, handler: h})
 }
 
-func (m handlerStorage) insert(endpoint string, entry handlerEntry) {
-	if m[endpoint] == nil {
-		m[endpoint] = list.New()
+func (hm handlerMapping) insert(endpoint string, entry handlerEntry) {
+	if hm[endpoint] == nil {
+		hm[endpoint] = new(internal.List[handlerEntry])
 	}
 
-	m[endpoint].PushBack(entry)
+	hm[endpoint].Insert(entry)
 }
 
 // forEndpoint returns handler what filters queries and execute correct handler.
-func (m handlerStorage) forEndpoint(endpoint string) Handler {
-	return func(teleCtx tele.Context, fsmCtx Context) error {
+func (m *Manager) forEndpoint(endpoint string) tele.HandlerFunc {
+	return func(teleCtx tele.Context) error {
+		fsmCtx := m.contextMaker(teleCtx, m.store)
+
 		state, err := fsmCtx.State()
 		if err != nil {
 			return &ErrHandlerState{Handler: endpoint, Err: err}
 		}
 
-		l := m[endpoint]
-
-		for e := l.Front(); e != nil; e = e.Next() {
-			h := e.Value.(handlerEntry)
-
-			if h.states.Has(state) || h.states.Has(AnyState) {
-				return h.handler(teleCtx, fsmCtx)
-			}
+		h, ok := m.handlers.find(endpoint, state)
+		if !ok {
+			return nil
 		}
-		return nil
+
+		// middlewares must be executed inside
+		// this handler for right work.
+		return h.handler(&wrapperContext{teleCtx, fsmCtx})
 	}
+}
+
+func (hm handlerMapping) find(endpoint string, state State) (handlerEntry, bool) {
+	l := hm[endpoint]
+
+	for e := l.Front(); e != nil; e = e.Next() {
+		h := e.Value
+
+		if h.states.Has(state) || h.states.Has(AnyState) {
+			return h, true
+		}
+	}
+
+	return handlerEntry{}, false
 }
 
 // ErrHandlerState indicates what manager gets error while tired
